+++ Commands/Information.cs
@@ -122,6 +122,17 @@ namespace JifBot.Commands
             await ReplyAsync("", false, embed.Build());
         }
 
+        [Command("invitelink")]
+        [Alias("link")]
+        [Remarks("-c-")]
+        [Summary("Provides a link which can be used should you want to spread Jif Bot to another server.")]
+        public async Task InviteLink()
+        {
+            var db = new BotBaseContext();
+            var config = db.Configuration.AsQueryable().Where(cfg => cfg.Name == Program.configName).First();
+            await ReplyAsync($"The following is a link to add me to another server. NOTE: You must have permissions on the server in order to add. Once on the server I must be given permission to send and delete messages, otherwise I will not work.\nhttps://discordapp.com/oauth2/authorize?client_id={config.Id}&scope=bot");
+        }
+
         [Command("define")]
         [Remarks("-c- word OR -c- word -m")]
         [Summary("Defines any word in the Oxford English dictionary. For multiple definitions, use -m at the end of the command.")]
@@ -510,72 +521,6 @@ namespace JifBot.Commands
                 var embed = ConstructEmbedInfo(await Context.Guild.GetUserAsync(Context.User.Id));
                 await ReplyAsync("", false, embed.Build());
             }
-        }
-
-        [Command("avatar")]
-        [Remarks("-c-, -c- @person1 @person2, -c- person1id person2id")]
-        [Summary("Gets the avatar for one or more users. Mention a user or provide their id to get their avatar, or do neither to get your own. To do more than 1 person, separate mentions/ids with spaces.")]
-        public async Task Avatar([Remainder] string ids = "")
-        {
-            await Context.Guild.DownloadUsersAsync();
-            var mention = Context.Message.MentionedUserIds;
-            if (mention.Count != 0)
-            {
-                foreach (ulong id in mention)
-                {
-                    var embed = new EmbedBuilder();
-                    IGuildUser user = Context.Guild.GetUserAsync(id).Result;
-                    string url = user.GetAvatarUrl();
-                    url = url.Remove(url.IndexOf("?size=128"));
-                    url = url + "?size=256";
-                    embed.ImageUrl = url;
-                    await ReplyAsync("", false, embed.Build());
-                }
-            }
-            else if (ids != "")
-            {
-                string[] idList = ids.Split(' ');
-                foreach (string id in idList)
-                {
-                    var embed = new EmbedBuilder();
-                    IGuildUser user = await Context.Guild.GetUserAsync(Convert.ToUInt64(id));
-                    string url = user.GetAvatarUrl();
-                    url = url.Remove(url.IndexOf("?size=128"));
-                    url = url + "?size=256";
-                    embed.ImageUrl = url;
-                    await ReplyAsync("", false, embed.Build());
-                }
-            }
-            else
-            {
-                var embed = new EmbedBuilder();
-                string url = Context.User.GetAvatarUrl();
-                url = url.Remove(url.IndexOf("?size=128"));
-                url = url + "?size=256";
-                embed.ImageUrl = url;
-                await ReplyAsync("", false, embed.Build());
-            }
-        }
-
-        [Command("funfact")]
-        [Remarks("-c-")]
-        [Alias("fact")]
-        [Summary("Provides a random fact.")]
-        public async Task beeFact([Remainder] string useless = "")
-        {
-            HttpClient client = new HttpClient();
-            client.BaseAddress = new Uri("https://useless-facts.sameerkumar.website");
-            HttpResponseMessage response = await client.GetAsync("/api");
-            HttpContent content = response.Content;
-            string stuff = await content.ReadAsStringAsync();
-            var json = JObject.Parse(stuff);
-            if ((string)json.SelectToken("Response") == "False")
-            {
-                await ReplyAsync("Error retrieving fact");
-                return;
-            }
-            string fact = (string)json.SelectToken("data");
-            await ReplyAsync(fact);
         }
 
         async Task<bool> RemoteFileExists(string url)
