+++ Commands/Utility.cs
@@ -7,11 +7,8 @@ using System.Net.Http;
 using System.Diagnostics;
 using Discord;
 using Discord.Commands;
-using Newtonsoft.Json;
 using JifBot.Models;
 using JIfBot;
-using System.Drawing;
-using System.IO;
 using System.Data;
 using Newtonsoft.Json.Linq;
 
@@ -19,114 +16,6 @@ namespace JifBot.Commands
 {
     public class Utility : ModuleBase
     {
-        List<Quote> quoteList = new List<Quote>();
-
-        [Command("invitelink")]
-        [Alias("link")]
-        [Remarks("-c-")]
-        [Summary("Provides a link which can be used should you want to spread Jif Bot to another server.")]
-        public async Task InviteLink()
-        {
-            var db = new BotBaseContext();
-            var config = db.Configuration.AsQueryable().Where(cfg => cfg.Name == Program.configName).First();
-            await ReplyAsync($"The following is a link to add me to another server. NOTE: You must have permissions on the server in order to add. Once on the server I must be given permission to send and delete messages, otherwise I will not work.\nhttps://discordapp.com/oauth2/authorize?client_id={config.Id}&scope=bot");
-        }
-
-        [Command("bigtext")]
-        [Remarks("-c- phrase, -c- phrase -d")]
-        [RequireBotPermission(ChannelPermission.ManageMessages)]
-        [Summary("Takes the user input for messages and turns it into large letters using emotes. If you end your command call with -d, it will delete your message calling the bot.")]
-        public async Task bigtext([Remainder]string orig)
-        {
-            if (orig.EndsWith("-d"))
-            {
-                orig = orig.Remove(orig.Length - 2);
-                await Context.Message.DeleteAsync();
-            }
-            string final = "";
-            orig = orig.ToLower();
-            for (int i = 0; i < orig.Length; i++)
-            {
-                final += getBig(orig[i]);
-                final += " ";
-            }
-            if (final.Length > 2000)
-                await ReplyAsync("This command does not support messages of that length, please shorten your message.");
-            else
-                await ReplyAsync(final);
-        }
-
-        [Command("tinytext")]
-        [Alias("smalltext")]
-        [Remarks("-c- phrase, -c- phrase -d")]
-        [RequireBotPermission(ChannelPermission.ManageMessages)]
-        [Summary("Takes the user input for messages and turns it into small letters. If you end your command call with -d, it will delete your message calling the bot.")]
-        public async Task tinytext([Remainder]string orig)
-        {
-            if (orig.EndsWith("-d"))
-            {
-                orig = orig.Remove(orig.Length - 2);
-                await Context.Message.DeleteAsync();
-            }
-            string final = "";
-            orig = orig.ToLower();
-            for (int i = 0; i < orig.Length; i++)
-            {
-                final += getSmol(orig[i]);
-            }
-            if (final.Length > 2000)
-                await ReplyAsync("This command does not support messages of that length, please shorten your message.");
-            else
-                await ReplyAsync(final);
-        }
-
-        [Command("widetext")]
-        [Remarks("-c- phrase, -c- phrase -d")]
-        [RequireBotPermission(ChannelPermission.ManageMessages)]
-        [Summary("Takes the user input for messages and turns it into a ＷＩＤＥ  ＢＯＩ. If you end your command call with -d, it will delete your message calling the bot.")]
-        public async Task WideText([Remainder] string message)
-        {
-            if (message.EndsWith("-d"))
-            {
-                message = message.Remove(message.Length - 2);
-                await Context.Message.DeleteAsync();
-            }
-            message = message.Replace(" ", "   ");
-            string alpha = "QWERTYUIOPASDFGHJKLÇZXCVBNMqwertyuiopasdfghjklçzxcvbnm,.-~+´«'0987654321!\"#$%&/()=?»*`^_:;";
-            string fullwidth = "ＱＷＥＲＴＹＵＩＯＰＡＳＤＦＧＨＪＫＬÇＺＸＣＶＢＮＭｑｗｅｒｔｙｕｉｏｐａｓｄｆｇｈｊｋｌçｚｘｃｖｂｎｍ,.－~ ´«＇０９８７６５４３２１！＂＃＄％＆／（）＝？»＊`＾＿：；";
-
-            for (int i = 0; i < alpha.Length; i++)
-            {
-                message = message.Replace(alpha[i], fullwidth[i]);
-            }
-            await ReplyAsync(message);
-        }
-
-        [Command("owo")]
-        [Alias("uwu")]
-        [Remarks("-c- phrase, -c- phrase -d")]
-        [RequireBotPermission(ChannelPermission.ManageMessages)]
-        [Summary("Takes the user input, and translates it into degenerate owo speak. If you end your command call with -d, it will delete your message calling the bot..")]
-        public async Task Owo([Remainder] string message)
-        {
-            if (message.EndsWith("-d"))
-            {
-                message = message.Remove(message.Length - 2);
-                await Context.Message.DeleteAsync();
-            }
-            string[] faces = new string[] { "(・ω・)", ";;w;;", "owo", "UwU", ">w<", "^w^" };
-            Random rnd = new Random();
-            message = Regex.Replace(message, @"(?:r|l)", "w");
-            message = Regex.Replace(message, @"(?:R|L)", "W");
-            message = Regex.Replace(message, @"n([aeiou])", @"ny$1");
-            message = Regex.Replace(message, @"N([aeiou])", @"Ny$1");
-            message = Regex.Replace(message, @"N([AEIOU])", @"NY$1");
-            message = Regex.Replace(message, @"ove", @"uv");
-            message = Regex.Replace(message, @"\!+", (match) => string.Format("{0}", " " + faces[rnd.Next(faces.Length)] + " "));
-
-            await ReplyAsync(message);
-        }
-
         [Command("timer")]
         [Remarks("-c- -h2 -m30 message, -c- 150 message")]
         [Summary("Sets a reminder to ping you after a certain amount of time has passed. A message can be specified along with the time to be printed back to you at the end of the timer. Times can be specified using any combination of -m[minutes], -h[hours], -d[days], and -w[weeks] anywhere in the message. Additionally, to set a quick timer for a number of minutes, just do -p-timer [minutes] message.")]
@@ -248,45 +137,6 @@ namespace JifBot.Commands
             html = html.Remove(0, html.IndexOf("?v=") + 3);
             html = html.Remove(html.IndexOf("\""));
             await ReplyAsync("https://www.youtube.com/watch?v=" + html);
-        }
-
-        [Command("mock")]
-        [Remarks("-c-, -c- message, -c- message -d")]
-        [RequireBotPermission(ChannelPermission.ManageMessages)]
-        [Summary("Mocks the text you provide it. If you end your command call with -d, it will delete your message calling the bot. If you do not specify any message, it will mock the most recent message sent in the text channel, and delete your command call.")]
-        public async Task Mock([Remainder] string words = "")
-        {
-            if (words == "")
-            {
-                await Context.Message.DeleteAsync();
-                var msg = Context.Channel.GetMessagesAsync(1).FlattenAsync().Result;
-                words = msg.ElementAt(0).Content;
-            }
-            else if (words.EndsWith("-d"))
-            {
-                words = words.Remove(words.Length - 2);
-                await Context.Message.DeleteAsync();
-            }
-            string end = string.Empty;
-            int i = 0;
-            foreach (char c in words)
-            {
-                if (c == ' ' || c == '"' || c == '.' || c == ',')
-                    end += c;
-                else if (i == 0)
-                {
-                    char temp = Char.ToLower(c);
-                    end += temp;
-                    i = 1;
-                }
-                else if (i == 1)
-                {
-                    char temp = Char.ToUpper(c);
-                    end += temp;
-                    i = 0;
-                }
-            }
-            await ReplyAsync(end);
         }
 
         [Command("8ball")]
@@ -414,83 +264,6 @@ namespace JifBot.Commands
 
         }
 
-        [Command("tiltycat")]
-        [Remarks("-c- degree")]
-        [Summary("Creates a cat at any angle you specify.\nSpecial thanks to Erik (Assisting#8734) for writing the program. Accessed via ```http://www.writeonlymedia.com/tilty_cat/(degree).png``` where (degree) is the desired angle.")]
-        public async Task TiltyCat(int degree, [Remainder] string useless = "")
-        {
-            Bitmap bmp = TiltyEmoji.tiltCat(degree);
-
-            using (MemoryStream ms = new MemoryStream())
-            {
-                bmp.Save(ms, System.Drawing.Imaging.ImageFormat.Png);
-                ms.Seek(0, SeekOrigin.Begin);
-                bmp.Dispose();
-                await Context.Channel.SendFileAsync(ms, "tiltycat.png");
-            }
-        }
-
-        [Command("joke")]
-        [Remarks("-c-")]
-        [Summary("Tells a joke.")]
-        public async Task Joke()
-        {
-            System.Net.Http.HttpClient client = new System.Net.Http.HttpClient();
-            string source = await client.GetStringAsync("http://www.rinkworks.com/jokes/random.cgi");
-            string ptr = "< div class='content'>";
-            source = source.Remove(0, source.IndexOf(ptr) + ptr.Length);
-            ptr = "</h2>";
-            source = source.Remove(0, source.IndexOf(ptr) + ptr.Length);
-            ptr = "</td><td class='ad'>";
-            source = source.Remove(source.IndexOf(ptr));
-            source = source.Replace("<p>", string.Empty);
-            source = source.Replace("</p>", "\n");
-            source = source.Replace("<ul>", "\n");
-            source = source.Replace("<li>", "\n");
-            source = source.Replace("<em>", "*");
-            source = source.Replace("</ul>", "\n");
-            source = source.Replace("</li>", "\n");
-            source = source.Replace(">/em>", "*");
-            await ReplyAsync(source);
-        }
-
-
-        [Command("inspire")]
-        [Remarks("-c-")]
-        [Summary("Gives an inspirational quote.")]
-        public async Task Inspire()
-        {
-            if (quoteList.Count == 0)
-            {
-                using (HttpClient client = new HttpClient())
-                {
-                    using (var response = await client.GetAsync("https://type.fit/api/quotes"))
-                    {
-                        string jsonResponse = await response.Content.ReadAsStringAsync();
-                        jsonResponse = jsonResponse.Replace("[", "{\"list\":[");
-                        jsonResponse = jsonResponse.Replace("]", "]}");
-                        try
-                        {
-                            QuoteResult result = JsonConvert.DeserializeObject<QuoteResult>(jsonResponse);
-                            quoteList = result.List;
-                            var target = result.List;
-                        }
-                        catch (Exception ex)
-                        {
-                            Console.WriteLine(ex);
-                        }
-                    }
-                }
-            }
-            int count = quoteList.Count;
-            Random rnd = new Random();
-            int num = rnd.Next(count);
-            Quote quote = quoteList[num];
-            if(quote.author == null)
-                quote.author = "Author Unknown";
-            await ReplyAsync("\"" + quote.text + "\"\n-" + quote.author);
-        }
-
         [Command("calculator")]
         [Remarks("-c- ( 5 + 7 ) / 2")]
         [Alias("calc", "math")]
@@ -543,127 +316,51 @@ namespace JifBot.Commands
             await ReplyAsync("https://strawpoll.com/" + (string)json.SelectToken("content_id"));
         }
 
-        public string getSmol(char orig)
+        [Command("avatar")]
+        [Remarks("-c-, -c- @person1 @person2, -c- person1id person2id")]
+        [Summary("Gets the avatar for one or more users. Mention a user or provide their id to get their avatar, or do neither to get your own. To do more than 1 person, separate mentions/ids with spaces.")]
+        public async Task Avatar([Remainder] string ids = "")
         {
-            switch (orig)
+            await Context.Guild.DownloadUsersAsync();
+            var mention = Context.Message.MentionedUserIds;
+            if (mention.Count != 0)
             {
-                case 'e':
-                    return "ᵉ";
-                case 't':
-                    return "ᵗ";
-                case 'a':
-                    return "ᵃ";
-                case 'o':
-                    return "ᵒ";
-                case 'i':
-                    return "ᶦ";
-                case 'n':
-                    return "ᶰ";
-                case 's':
-                    return "ˢ";
-                case 'r':
-                    return "ʳ";
-                case 'h':
-                    return "ʰ";
-                case 'd':
-                    return "ᵈ";
-                case 'l':
-                    return "ᶫ";
-                case 'u':
-                    return "ᵘ";
-                case 'c':
-                    return "ᶜ";
-                case 'm':
-                    return "ᵐ";
-                case 'f':
-                    return "ᶠ";
-                case 'y':
-                    return "ʸ";
-                case 'w':
-                    return "ʷ";
-                case 'g':
-                    return "ᵍ";
-                case 'p':
-                    return "ᵖ";
-                case 'b':
-                    return "ᵇ";
-                case 'v':
-                    return "ᵛ";
-                case 'k':
-                    return "ᵏ";
-                case 'x':
-                    return "ˣ";
-                case 'q':
-                    return "ᑫ";
-                case 'j':
-                    return "ʲ";
-                case 'z':
-                    return "ᶻ";
+                foreach (ulong id in mention)
+                {
+                    var embed = new EmbedBuilder();
+                    IGuildUser user = Context.Guild.GetUserAsync(id).Result;
+                    string url = user.GetAvatarUrl();
+                    url = url.Remove(url.IndexOf("?size=128"));
+                    url = url + "?size=256";
+                    embed.ImageUrl = url;
+                    await ReplyAsync("", false, embed.Build());
+                }
             }
-            return Convert.ToString(orig);
+            else if (ids != "")
+            {
+                string[] idList = ids.Split(' ');
+                foreach (string id in idList)
+                {
+                    var embed = new EmbedBuilder();
+                    IGuildUser user = await Context.Guild.GetUserAsync(Convert.ToUInt64(id));
+                    string url = user.GetAvatarUrl();
+                    url = url.Remove(url.IndexOf("?size=128"));
+                    url = url + "?size=256";
+                    embed.ImageUrl = url;
+                    await ReplyAsync("", false, embed.Build());
+                }
+            }
+            else
+            {
+                var embed = new EmbedBuilder();
+                string url = Context.User.GetAvatarUrl();
+                url = url.Remove(url.IndexOf("?size=128"));
+                url = url + "?size=256";
+                embed.ImageUrl = url;
+                await ReplyAsync("", false, embed.Build());
+            }
         }
 
-        public string getBig(char orig)
-        {
-            switch (orig)
-            {
-                case 'e':
-                    return "🇪";
-                case 't':
-                    return "🇹";
-                case 'a':
-                    return "🇦";
-                case 'o':
-                    return "🇴";
-                case 'i':
-                    return "🇮";
-                case 'n':
-                    return "🇳";
-                case 's':
-                    return "🇸";
-                case 'r':
-                    return "🇷";
-                case 'h':
-                    return "🇭";
-                case 'd':
-                    return "🇩";
-                case 'l':
-                    return "🇱";
-                case 'u':
-                    return "🇺";
-                case 'c':
-                    return "🇨";
-                case 'm':
-                    return "🇲";
-                case 'f':
-                    return "🇫";
-                case 'y':
-                    return "🇾";
-                case 'w':
-                    return "🇼";
-                case 'g':
-                    return "🇬";
-                case 'p':
-                    return "🇵";
-                case 'b':
-                    return "🅱";
-                case 'v':
-                    return "🇻";
-                case 'k':
-                    return "🇰";
-                case 'x':
-                    return "🇽";
-                case 'q':
-                    return "🇶";
-                case 'j':
-                    return "🇯";
-                case 'z':
-                    return "🇿";
-                case ' ':
-                    return "  ";
-            }
-            return Convert.ToString(orig);
-        }
         string formatMinutesToString(int minutes)
         {
             string format = "";
@@ -718,58 +415,5 @@ namespace JifBot.Commands
     class UrbanDictionaryResult
     {
         public List<UrbanDictionaryDefinition> List { get; set; }
-    }
-
-    class Quote
-    {
-        public string text { get; set; }
-        public string author { get; set; }
-    }
-
-    class QuoteResult
-    {
-        public List<Quote> List { get; set; }
-    }
-
-    class TiltyEmoji
-    {
-        private const string TILTY_CAT = "Media/tiltycat.png";
-        private const int OUTPUT_WIDTH = 128;
-        private const int OUTPUT_HEIGHT = 128;
-
-        public static Bitmap tiltCat(int degreeCount)
-        {
-            return openAndRotate(TILTY_CAT, degreeCount);
-        }
-
-        private static Bitmap openAndRotate(string filePath, int degreeCount)
-        {
-            try
-            {
-                Bitmap bmp = (Bitmap)Bitmap.FromFile(filePath);
-                return RotateImage(bmp, degreeCount);
-            }
-            catch (FileNotFoundException ex)
-            {
-                Console.Out.WriteLine($"Couldn't find image at file path {filePath}");
-                throw ex;
-            }
-        }
-
-        private static Bitmap RotateImage(Bitmap bmp, int angle)
-        {
-            float height = bmp.Height;
-            float width = bmp.Width;
-            int hypotenuse = System.Convert.ToInt32(System.Math.Floor(Math.Sqrt(height * height + width * width)));
-            Bitmap rotatedImage = new Bitmap(hypotenuse, hypotenuse);
-            using (Graphics g = Graphics.FromImage(rotatedImage))
-            {
-                g.TranslateTransform((float)rotatedImage.Width / 2, (float)rotatedImage.Height / 2); //set the rotation point as the center into the matrix
-                g.RotateTransform(angle); //rotate
-                g.TranslateTransform(-(float)rotatedImage.Width / 2, -(float)rotatedImage.Height / 2); //restore rotation point into the matrix
-                g.DrawImage(bmp, (hypotenuse - OUTPUT_WIDTH) / 2, (hypotenuse - OUTPUT_HEIGHT) / 2, OUTPUT_WIDTH, OUTPUT_HEIGHT);
-            }
-            return rotatedImage;
-        }
     }
 }
